move generic functionality into shared
